CRD generation (#68)
Generate CRD instead of installing it from the code on startup. CRD installation is delegated to Helm or manual installations. Which should work the same as it does now if you install the operator with Helm. It will install the CRD on first installation, further updates will have to be done outside Helm.